+++ Makefile.tools
@@ -323,15 +323,16 @@ tools_btgatt_client_SOURCES = tools/btgatt-client.c src/uuid-helper.c
 tools_btgatt_client_LDADD = src/libshared-mainloop.la \
 						lib/libbluetooth-internal.la
 
-tools_pk_btgatt_hack_SOURCES = tools/pk-btgatt-hack.c src/uuid-helper.c
+tools_pk_btgatt_hack_SOURCES = tools/pk-btgatt-hack.c \
+                               tools/pk-btgatt-tcpip.c \
+                               src/uuid-helper.c
+
 tools_pk_btgatt_hack_LDADD   = src/libshared-mainloop.la \
-						       lib/libbluetooth-internal.la
+                               lib/libbluetooth-internal.la
 
 tools_btgatt_server_SOURCES = tools/btgatt-server.c src/uuid-helper.c
 tools_btgatt_server_LDADD = src/libshared-mainloop.la \
 						lib/libbluetooth-internal.la
-
-bogan_SOURCES = tools/bogan.cpp
 
 EXTRA_DIST += tools/bdaddr.1
 endif
